update weebhook activity and kas

+++ pages/admin/mahasiswa/index.vue
@@ -2,25 +2,22 @@
 import {
   mdiMonitorCellphone,
   mdiTableBorder,
-  mdiTableOff,
-  mdiEye,
+
   mdiTrashCan,
-  mdiGithub,
+
 } from "@mdi/js";
 import SectionMain from "@/components/SectionMain.vue";
 import NotificationBar from "@/components/NotificationBar.vue";
-import TableSampleClients from "@/components/TableSampleClients.vue";
+
 import CardBox from "@/components/CardBox.vue";
-// import LayoutAuthenticated from "@/layouts/LayoutAuthenticated.vue";
+
 import SectionTitleLineWithButton from "@/components/SectionTitleLineWithButton.vue";
 import BaseButton from "@/components/BaseButton.vue";
 import BaseButtons from "@/components/BaseButtons.vue";
-// import BaseButtons from "@/components/BaseButtons.vue";
-import CardBoxComponentEmpty from "@/components/CardBoxComponentEmpty.vue";
 
 
-const isModalActive = ref(false);
-const isModalDangerActive = ref(false);
+
+
 const mahasiswa = ref([]);
 const loading = ref(false)
 const supabase = useSupabaseClient();
@@ -34,11 +31,18 @@ const getMahasiswa = async ()=>{
     loading.value = true
 }
 const deleteMahasiswa = async (id)=>{
+  if(confirm('Yakin ingin menghapus data?')){
+
     const { data, error } = await supabase
     .from('mahasiswa')
     .delete()
     .match({ id: id })
     getMahasiswa()
+  }
+  else{
+    alert('data tidak jadi dihapus')
+  
+  }
 }
 onMounted(()=>{
     getMahasiswa()
